Move fileExistsSync to utils
